FIX: Allow manual scaling

+++ mne/viz/tests/test_decoding.py
@@ -79,7 +79,7 @@ def test_gat_plot_times():
     gat.plot_times(gat.train_times['times_'])
     # test multiple colors
     n_times = len(gat.train_times['times_'])
-    colors = np.tile(['r', 'g', 'b'], np.ceil(n_times / 3))[:n_times]
+    colors = np.tile(['r', 'g', 'b'], int(np.ceil(n_times / 3)))[:n_times]
     gat.plot_times(gat.train_times['times_'], color=colors)
     # test invalid time point
     assert_raises(ValueError, gat.plot_times, -1.)
